issue-6996: fastshard - extracted StorageGroupFactory away from the shard implementation library
