+++ cloud/filestore/libs/storage/fastshard/iface/fs.h
@@ -98,19 +98,6 @@ struct IFileSystemShard
 
 ////////////////////////////////////////////////////////////////////////////////
 
-struct IFileSystemShardFactory
-{
-    virtual ~IFileSystemShardFactory() = default;
-
-    virtual IFileSystemShardPtr CreateShard(
-        const TString& fileSystemId,
-        const NProtoPrivate::TFastShardConfig& config,
-        ui32 shardNo,
-        ui64 generation) = 0;
-};
-
-////////////////////////////////////////////////////////////////////////////////
-
 IFileSystemShardPtr CreateFileSystemShardStub();
 
 }   // namespace NCloud::NFileStore::NStorage::NFastShard
