+++ cloud/filestore/libs/storage/fastshard/impl/bench/ya.make
@@ -15,6 +15,7 @@ PEERDIR(
     cloud/filestore/libs/service
     cloud/filestore/libs/storage/fastshard/iface
     cloud/filestore/libs/storage/fastshard/impl/hash_table_index
+    cloud/filestore/libs/storage/fastshard/sn/factory
     cloud/filestore/libs/storage/fastshard/sn/quorum
     cloud/filestore/private/api/protos
