+++ cloud/filestore/libs/storage/fastshard/impl/hash_table_index/shard.cpp
@@ -12,9 +12,8 @@
 #include <cloud/filestore/libs/storage/fastshard/impl/model/persistent_bitmap.h>
 #include <cloud/filestore/libs/storage/fastshard/impl/model/persistent_hash_table.h>
 #include <cloud/filestore/libs/storage/fastshard/sn/client/client.h>
+#include <cloud/filestore/libs/storage/fastshard/sn/factory/group_factory.h>
 #include <cloud/filestore/libs/storage/fastshard/sn/quorum/storage_group.h>
-#include <cloud/filestore/libs/storage/fastshard/sn/quorum/storage_group_helpers.h>
-#include <cloud/filestore/libs/storage/fastshard/sn/quorum/storage_group_quorum.h>
 #include <cloud/filestore/libs/storage/model/utils.h>
 #include <cloud/filestore/private/api/protos/tablet.pb.h>
 
@@ -2145,59 +2144,7 @@ public:
     }
 };
 
-////////////////////////////////////////////////////////////////////////////////
-
-struct TStorageGroupFactory: IStorageGroupFactory
-{
-    IStorageGroupPtr MakeStorageGroup(
-        const NProtoPrivate::TPersistentFastShardConfig& config,
-        ui64 generation) override
-    {
-        TVector<TStorageDevice> devices;
-        const auto& sg = config.GetStorageGroups(0);
-        for (const auto& d: sg.GetDevices()) {
-            devices.push_back({
-                .Node = CreateStorageNodeClient(d.GetHost(), d.GetPort()),
-                .DeviceUUID = d.GetDeviceId(),
-            });
-        }
-
-        TStorageGroupConfig groupConfig;
-        // TODO(#5894): set client id
-        groupConfig.AcquireGeneration = generation;
-        if (config.GetRetryTotalTimeoutMs()) {
-            groupConfig.RetryPolicy.TotalTimeout =
-                TDuration::MilliSeconds(config.GetRetryTotalTimeoutMs());
-        }
-        if (config.GetRetryBackoffIncrementMs()) {
-            groupConfig.RetryPolicy.BackoffIncrement =
-                TDuration::MilliSeconds(config.GetRetryBackoffIncrementMs());
-        }
-
-        groupConfig.JournalRestoreEnabled = config.GetJournalRestoreEnabled();
-
-        if (sg.GetType() == NProtoPrivate::TStorageGroup::E_SG_QUORUM_MIRROR) {
-            return CreateQuorumMirroredStorageGroup(
-                std::move(groupConfig),
-                std::move(devices),
-                CreateFiberTimer());
-        }
-
-        return CreateNaiveMirroredStorageGroup(
-            std::move(groupConfig),
-            std::move(devices),
-            CreateFiberTimer());
-    }
-};
-
 }   // namespace
-
-////////////////////////////////////////////////////////////////////////////////
-
-IStorageGroupFactoryPtr CreateStorageGroupFactory()
-{
-    return std::make_shared<TStorageGroupFactory>();
-}
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -2234,20 +2181,6 @@ IFileSystemShardPtr CreateHashTableIndexFileSystemShard(
         shardNo,
         generation,
         std::move(storageGroupFactory),
-        config);
-}
-
-IFileSystemShardPtr CreateHashTableIndexFileSystemShard(
-    TString fileSystemId,
-    ui32 shardNo,
-    ui64 generation,
-    const NProtoPrivate::TPersistentFastShardConfig& config)
-{
-    return std::make_shared<THashTableIndexFileSystemShard>(
-        std::move(fileSystemId),
-        shardNo,
-        generation,
-        CreateStorageGroupFactory(),
         config);
 }
