+++ cloud/filestore/libs/storage/fastshard/sn/factory/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+SRCS(
+    group_factory.cpp
+)
+
+PEERDIR(
+    cloud/filestore/libs/storage/fastshard/sn/quorum
+
+    cloud/filestore/private/api/protos
+)
+
+END()
